Abstract VideoCapture from cv2
The ordinary VideoCapture from cv2 is very unpythonic and very
unintuitive to python users. As such a new Video class is added that
creates a more pythonic interface to working with Videos. A lot of the
properties ordinarily gotten with VideoCapture.get() are now properties
in the python sense.
The Video object is also usable as an iterator to allow for going
through a whole file in succession.