Introduce new way for creating segmentations.
Segmentations are now created by inheriting from BaseSegmentation and
defining a series of dataclass parameters (.params) inherited from BaseParam.
The dataclasses provide a .widget() for creating a dictionary containing
the widgets to be added.
The BaseSegmentation .create_gui() method then is able to arrange these
into a grid.

Changes to widgets.py are necessary because Segmentation .widgets no
longer are structured in the same way as previously.